Rename MultiplyAccumulate to ArithmeticOperations
